+++ unijobs-core/src/main/java/core/service/FetchService.java
@@ -1,6 +1,7 @@
 package core.service;
 
 import core.model.Skill;
+import core.model.UniUser;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,4 +10,6 @@ public interface FetchService {
     List<Skill> getAllSkill();
 
     Skill findSkill(Long id);
+
+    List<UniUser> getAllUsers();
 }
